Crate warheadhateus

AWS Authorization Header Generation (AWS Signature Version 4)

After setting up the AWSAuth struct, calling auth_header will return a String similar to the following:

AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE/20130524/us-east-1/s3/aws4_request,\
SignedHeaders=host;range;x-amz-content-sha256;x-amz-date,\
Signature=f0e8bdb87c964420e857bd35b5d6ed310bd44f0170aba48dd91039c6036bdb41

Currently this is implemented for S3 single and multi chunk mode and passes all tests.

Single Chunk

Multiple Chunks

Examples

Amazon S3 Single Chunk Mode

let mut auth = AWSAuth::new(URL_1).unwrap_or_else(|e| fail(e));
let payload_hash = hashed_data(None).unwrap_or_else(|e| fail(e));
let scope_date = UTC.datetime_from_str(SCOPE_DATE, DATE_TIME_FMT).unwrap_or_else(|e| fail(e));
auth.set_request_type(HttpRequestMethod::GET);
auth.set_payload_hash(&payload_hash);
auth.set_date(scope_date);
auth.set_service(Service::S3);
auth.set_access_key_id(ACCESS_KEY_ID);
auth.set_secret_access_key(SECRET_ACCESS_KEY);
auth.set_region(Region::UsEast1);
auth.add_header("HOST", HOST);
auth.add_header("x-amz-content-sha256", &payload_hash);
auth.add_header("x-amz-date", SCOPE_DATE);
auth.add_header("Range", "bytes=0-9");

let ah = auth.auth_header().unwrap_or_else(|e| fail(e));
assert!(ah == AWS_TEST_1);

Structs

AWSAuth
Amazon Web Service Authorization Header struct

Enums

AWSAuthError
Authentication Error Types

HttpRequestMethod
See RFC7231

Mode
Amazon S3 Mode of Operation

Region
AWS Region

SAM
Singing Algorithm Moniker

Service
AWS Services

Functions

hashed_data
Hash the given data (or an empty string) with SHA256.

Type Definitions

AWSAuthResult
AWS Authentication Header Generation Result.