Struct warheadhateus::AWSAuth

pub struct AWSAuth {
    // some fields omitted
}

Amazon Web Service Authorization Header struct

Methods

impl AWSAuth

fn new(url: &str) -> Result<AWSAuth, AWSAuthError>

Create a new AWSAuth struct.

fn add_header(&mut self, key: &str, val: &str) -> &mut AWSAuth

Add a header field to the AWSAuth struct.

fn set_access_key_id(&mut self, access_key_id: &str) -> &mut AWSAuth

Set the AWS Access Key ID (this is supplied by Amazon).

fn set_date(&mut self, date: DateTime<UTC>) -> &mut AWSAuth

Set the scope date for the auth request. Request are valid for 7 days from the given date. * Note that this date doesn't have to match the x-amz-date or date header values.

fn set_payload_hash(&mut self, payload_hash: &str) -> &mut AWSAuth

Set the payload hash. This should be the hash of any request payload, or the hash of the empty string ("") if there is no payload.

fn set_region(&mut self, region: Region) -> &mut AWSAuth

Set the region the request is scoped for.

fn set_request_type(&mut self, verb: HttpRequestMethod) -> &mut AWSAuth

Set the request verb (i.e. GET, PUT, POST, DELETE)

fn set_secret_access_key(&mut self, secret_access_key: &str) -> &mut AWSAuth

Set the AWS Secret Access Key (this is supplied by Amazon).

fn set_service(&mut self, service: Service) -> &mut AWSAuth

Set the AWS services this request should be constructed for (i.e. s3)

fn auth_header(&self) -> AWSAuthResult

Create the AWS S3 Authorization HTTP header.