Struct warheadhateus::AWSAuth

pub struct AWSAuth {
    // some fields omitted
}

Amazon Web Service Authorization Header struct

Methods

impl AWSAuth

fn new(url: &str) -> Result<AWSAuth, AWSAuthError>

Create a new AWSAuth struct.

Examples

use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(auth) => {
        // Do fun stuff with the auth struct
    }
    Err(_) => {
        // Failure
    }
}

fn add_header(&mut self, key: &str, val: &str) -> &mut AWSAuth

Add a header key/value pair.

Examples

use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.add_header("Host", "s3.amazonaws.com");
    }
    Err(_) => {
        // Failure
    }
}

fn set_access_key_id(&mut self, access_key_id: &str) -> &mut AWSAuth

Set the AWS Access Key ID (this is supplied by Amazon).

Examples

use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_access_key_id("AKIAIOSFODNN7EXAMPLE");
    }
    Err(_) => {
        // Failure
    }
}

fn set_chunk_size(&mut self, chunk_size: usize) -> &mut AWSAuth

Set the chunk size.

Note this is only used when Mode::Chunked is selected.

Chunk size must be at least 8 KB. We recommend a chunk size of a least 64 KB for better performance. This chunk size applies to all CHUNK_SIG_3 except the last one. The last chunk you send can be smaller than 8 KB. If your payload is small and can fit in one chunk, then it can be smaller than the 8 KB.

Examples

use warheadhateus::{AWSAuth, Mode};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_mode(Mode::Chunked);
        auth.set_chunk_size(65536);
    }
    Err(_) => {
        // Failure
    }
}

fn set_date(&mut self, date: DateTime<UTC>) -> &mut AWSAuth

Set the scope date for the auth request.

Note that this date doesn't have to match the x-amz-date or date header values.

Requests are valid for 7 days from the given scope date.

Examples

use chrono::UTC;
use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_date(UTC::now());
    }
    Err(_) => {
        // Failure
    }
}

fn set_mode(&mut self, mode: Mode) -> &mut AWSAuth

Set the mode of operation.

Examples

use warheadhateus::{AWSAuth, Mode};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_mode(Mode::Chunked);
    }
    Err(_) => {
        // Failure
    }
}

fn set_payload_hash(&mut self, payload_hash: &str) -> &mut AWSAuth

Set the payload hash.

This should be the SHA-256 hash of any request payload, or the hash of the empty string ("") if there is no payload. There is a hashed_data function in this library that will do this for you, but you are free to use any SHA-256 implementation.

Examples

use warheadhateus::{AWSAuth, hashed_data};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        if let Ok(hd) = hashed_data(Some(b"This is a test")) {
            auth.set_payload_hash(&hd);
        }
    }
    Err(_) => {
        // Failure
    }
}

fn set_region(&mut self, region: Region) -> &mut AWSAuth

Set the region the request is scoped for.

Examples

use warheadhateus::{AWSAuth, Region};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_region(Region::UsWest1);
    }
    Err(_) => {
        // Failure
    }
}

fn set_request_type(&mut self, verb: HttpRequestMethod) -> &mut AWSAuth

Set the request verb (i.e. GET, PUT, POST, DELETE)

Examples

use warheadhateus::{AWSAuth, HttpRequestMethod};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_request_type(HttpRequestMethod::PUT);
    }
    Err(_) => {
        // Failure
    }
}

fn set_secret_access_key(&mut self, secret_access_key: &str) -> &mut AWSAuth

Set the AWS Secret Access Key (this is supplied by Amazon).

Examples

use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_secret_access_key("wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY");
    }
    Err(_) => {
        // Failure
    }
}

fn set_sam(&mut self, sam: SAM) -> &mut AWSAuth

Set the Signing Algorithm Moniker (SAM).

Note that the default is usually acceptable here, only chunked requests need to change it.

Examples

use warheadhateus::{AWSAuth, SAM};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_sam(SAM::AWS4HMACSHA256PAYLOAD);
    }
    Err(_) => {
        // Failure
    }
}

fn set_seed(&mut self, seed: bool) -> &mut AWSAuth

Set the seed flag.

Examples

use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_seed(true);
    }
    Err(_) => {
        // Failure
    }
}

fn set_service(&mut self, service: Service) -> &mut AWSAuth

Set the AWS services this request should be constructed for (i.e. S3)

Examples

use warheadhateus::{AWSAuth, Service};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_service(Service::S3);
    }
    Err(_) => {
        // Failure
    }
}

fn set_version(&mut self, version: SigningVersion) -> &mut AWSAuth

Set the Signing Version (2 or 4)

Examples

use warheadhateus::{AWSAuth, SigningVersion};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_version(SigningVersion::Two);
    }
    Err(_) => {
        // Failure
    }
}

fn signature(&self) -> AWSAuthResult

Generate the signature.

Examples

use std::io::{self, Write};
use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        if let Ok(sig) = auth.signature() {
            writeln!(io::stdout(), "{}", sig).expect("Unable to write to stdout!");
        }
    }
    Err(_) => {
        // Failure
    }
}

fn auth_header(&self) -> AWSAuthResult

Create the AWS Authorization HTTP header.

Examples

use std::io::{self, Write};
use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        if let Ok(ah) = auth.auth_header() {
            writeln!(io::stdout(), "{}", ah).expect("Unable to write to stdout!");
        }
    }
    Err(_) => {
        // Failure
    }
}

fn query_string(&self) -> AWSAuthResult

Create the AWS Authorization Query String.

Examples

use std::io::{self, Write};
use warheadhateus::AWSAuth;

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        if let Ok(qs) = auth.query_string() {
            writeln!(io::stdout(), "{}", qs).expect("Unable to write to stdout!");
        }
    }
    Err(_) => {
        // Failure
    }
}

fn seed_signature(&self) -> AWSAuthResult

Generate the seed signature for chunked mode.

Examples

use std::io::{self, Write};
use warheadhateus::{AWSAuth, Mode};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_mode(Mode::Chunked);
        auth.set_seed(true);
        auth.set_chunk_size(65536);
        if let Ok(ss) = auth.seed_signature() {
            writeln!(io::stdout(), "{}", ss).expect("Unable to write to stdout!");
        }
    }
    Err(_) => {
        // Failure
    }
}

fn chunk_signature(&self, previous_signature: &str, chunk: &[u8]) -> AWSAuthResult

Generate the chunk signature for a given chunk.

Examples

use std::io::{self, Write};
use warheadhateus::{AWSAuth, Mode};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_mode(Mode::Chunked);
        auth.set_chunk_size(65536);
        if let Ok(ss) = auth.seed_signature() {
            if let Ok(cs) = auth.chunk_signature(&ss, &[0; 100]) {
                writeln!(io::stdout(), "{}", cs).expect("Unable to write to stdout!");
            }
        }
    }
    Err(_) => {
        // Failure
    }
}

fn chunk_body(&self, chunk_signature: &str, chunk: &[u8]) -> Result<Vec<u8>, AWSAuthError>

Generate the chunk body for a given chunk.

Examples

use std::io::{self, Write};
use warheadhateus::{AWSAuth, Mode};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_mode(Mode::Chunked);
        auth.set_chunk_size(65536);
        if let Ok(ss) = auth.seed_signature() {
            let chunk = [0; 100];
            if let Ok(cs) = auth.chunk_signature(&ss, &chunk) {
                if let Ok(cb) = auth.chunk_body(&cs, &chunk) {
                    let l = cb.len();
                    writeln!(io::stdout(), "{}", l).expect("Unable to write to stdout!");
                }
            }
        }
    }
    Err(_) => {
        // Failure
    }
}

fn content_length(&self, payload_size: usize) -> Result<usize, AWSAuthError>

Generate the value for the 'Content-Length' header when using chunked mode.

This function can be used for the following requirement from Amazon for pre-computing the Content-Length header.

When transferring data in a series of chunks, you must use the Content-Length HTTP header to explicitly specify the total content length (object length in bytes plus metadata in each chunk). This will require you to pre-compute the total length of the payload including the metadata you will send in each chunk before starting your request. The x-amz-decoded-content-length header will contain the size of the object length in bytes.

Examples

use std::io::{self, Write};
use warheadhateus::{AWSAuth, Mode};

match AWSAuth::new("https://s3.amazonaws.com/examplebucket/test.txt") {
    Ok(mut auth) => {
        auth.set_mode(Mode::Chunked);
        auth.set_chunk_size(65536);
        if let Ok(cl) = auth.content_length(100000) {
            writeln!(io::stdout(), "{}", cl).expect("Unable to write to stdout!");
        }
    }
    Err(_) => {
        // Failure
    }
}